Refactor airport app structure and add image upload functionality
- Renamed common app to base app.
- Introduced AirplaneImageSerializer for handling airplane image uploads.
- Updated AirplaneViewSet to include a new action for image uploads.
